# src/AppForSEII2526.Web/OpenAPIs/swagger1.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "AppForMovies.API",
    "description": "This API provides services for renting and purchasing movies",
    "contact": {
      "name": "Software Engineering II Team",
      "email": "isii@on.uclm.es"
    },
    "license": {
      "name": "MIT License",
      "url": "https://opensource.org/license/mit/"
    },
    "version": "v1"
  },
  "paths": {
...
    "/api/Plan/GetPlan": {
      "get": {
        "tags": [
          "Plan"
        ],
        "operationId": "GetPlan",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetPlanDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetPlanDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetPlanDTO"
                }
              }
            }
          },
          "404": {
            "description": "Not Found",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ProblemDetails"
                }
              }
            }
          }
        }
      }
    },
    "/api/Plan/CreatePlan": {
      "post": {
        "tags": [
          "Plan"
        ],
        "operationId": "CreatePlan",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlanDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlanDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePlanDTO"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/GetPlanDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetPlanDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/GetPlanDTO"
                }
              }
            }
          },
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      }
    },
    "/api/Purchase/GetPurchase": {
      "get": {
        "tags": [
          "Purchase"
        ],
        "operationId": "GetPurchase",
        "parameters": [
          {
            "name": "id",
            "in": "query",
            "schema": {
              "type": "integer",
              "format": "int32"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PurchaseDTO"
                  }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PurchaseDTO"
                  }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/PurchaseDTO"
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/Purchase/CreatePurchase": {
      "post": {
        "tags": [
          "Purchase"
        ],
        "operationId": "CreatePurchase",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePurchaseDTO"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePurchaseDTO"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CreatePurchaseDTO"
              }
            }
          }
        },
        "responses": {
          "400": {
            "description": "Bad Request",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidationProblemDetails"
                }
              }
            }
          },
          "409": {
            "description": "Conflict",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "string"
                }
              },
              "application/json": {
                "schema": {
                  "type": "string"
                }
              },
              "text/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          },
          "201": {
            "description": "Created",
            "content": {
              "text/plain": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseDTO"
                }
              },
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseDTO"
                }
              },
              "text/json": {
                "schema": {
                  "$ref": "#/components/schemas/PurchaseDTO"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
      "ApplicationUser": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "userName": {
            "type": "string",
            "nullable": true
          },
          "normalizedUserName": {
            "type": "string",
            "nullable": true
          },
          "email": {
            "type": "string",
            "nullable": true
          },
          "normalizedEmail": {
            "type": "string",
            "nullable": true
          },
          "emailConfirmed": {
            "type": "boolean"
          },
          "passwordHash": {
            "type": "string",
            "nullable": true
          },
          "securityStamp": {
            "type": "string",
            "nullable": true
          },
          "concurrencyStamp": {
            "type": "string",
            "nullable": true
          },
          "phoneNumber": {
            "type": "string",
            "nullable": true
          },
          "phoneNumberConfirmed": {
            "type": "boolean"
          },
          "twoFactorEnabled": {
            "type": "boolean"
          },
          "lockoutEnd": {
            "type": "string",
            "format": "date-time",
            "nullable": true
          },
          "lockoutEnabled": {
            "type": "boolean"
          },
          "accessFailedCount": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "surname": {
            "type": "string",
            "nullable": true
          },
          "paymentMethods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentMethod"
            },
            "nullable": true
          },
          "incidents": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Incident"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "Class": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "capacity": {
            "type": "integer",
            "format": "int32"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "planItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlanItem"
            },
            "nullable": true
          },
          "typeItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/TypeItem"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "ClassInPlanDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "maxLength": 50,
            "minLength": 3,
            "type": "string",
            "nullable": true
          },
          "typeItemNames": {
            "type": "array",
            "items": {
              "type": "string"
            },
            "nullable": true
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "goal": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "CreatePlanDTO": {
        "required": [
          "name",
          "selectedPaymentMethodId",
          "weeks"
        ],
        "type": "object",
        "properties": {
          "selectedClasses": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClassInPlanDTO"
            },
            "nullable": true
          },
          "name": {
            "maxLength": 50,
            "minLength": 3,
            "type": "string"
          },
          "description": {
            "maxLength": 500,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "weeks": {
            "maximum": 52,
            "minimum": 1,
            "type": "integer",
            "format": "int32"
          },
          "healthIssues": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "availablePaymentMethods": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PaymentMethodDTO"
            },
            "nullable": true
          },
          "selectedPaymentMethodId": {
            "type": "integer",
            "format": "int32"
          }
        },
        "additionalProperties": false
      },
      "CreatePurchaseDTO": {
        "type": "object",
        "properties": {
          "city": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "street": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "total_price": {
            "type": "number",
            "format": "double"
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "purchaseItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PurchaseItemsDTO"
            },
            "nullable": true
          },
          "paymentMethod": {
            "$ref": "#/components/schemas/PaymentMethod"
          }
        },
        "additionalProperties": false
      },
      "GetPlanDTO": {
        "type": "object",
        "properties": {
          "userName": {
            "maxLength": 50,
            "minLength": 3,
            "type": "string",
            "nullable": true
          },
          "userSurname": {
            "maxLength": 50,
            "minLength": 3,
            "type": "string",
            "nullable": true
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "totalPrice": {
            "type": "number",
            "format": "double"
          },
          "name": {
            "maxLength": 50,
            "minLength": 3,
            "type": "string",
            "nullable": true
          },
          "description": {
            "maxLength": 500,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "weeks": {
            "type": "integer",
            "format": "int32"
          },
          "healthIssues": {
            "maxLength": 200,
            "minLength": 0,
            "type": "string",
            "nullable": true
          },
          "classes": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/ClassInPlanDTO"
            },
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Incident": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "dateOfldentification": {
            "type": "string",
            "format": "date-time"
          },
          "exercise": {
            "type": "string",
            "nullable": true
          },
          "incidentItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/IncidentItem"
            },
            "nullable": true
          },
          "user": {
            "$ref": "#/components/schemas/ApplicationUser"
          },
          "incidentState": {
            "$ref": "#/components/schemas/IncidentState"
          }
        },
        "additionalProperties": false
      },
      "IncidentItem": {
        "type": "object",
        "properties": {
          "incidentId": {
            "type": "integer",
            "format": "int32"
          },
          "itemForExerciseId": {
            "type": "integer",
            "format": "int32"
          },
          "incident": {
            "$ref": "#/components/schemas/Incident"
          },
          "itemForExercise": {
            "$ref": "#/components/schemas/ItemForExercise"
          },
          "incidentPriority": {
            "$ref": "#/components/schemas/IncidentPriority"
          }
        },
        "additionalProperties": false
      },
...
      "PaymentMethod": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "plans": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Plan"
            },
            "nullable": true
          },
          "purchases": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/Purchase"
            },
            "nullable": true
          },
          "user": {
            "$ref": "#/components/schemas/ApplicationUser"
          }
        },
        "additionalProperties": false
      },
      "PaymentMethodDTO": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "type": {
            "type": "string",
            "nullable": true
          },
          "displayInfo": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "Plan": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "createdDate": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "healthIssues": {
            "type": "string",
            "nullable": true
          },
          "name": {
            "type": "string",
            "nullable": true
          },
          "totalPrice": {
            "type": "number",
            "format": "double"
          },
          "weeks": {
            "type": "integer",
            "format": "int32"
          },
          "planItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PlanItem"
            },
            "nullable": true
          },
          "paymentMethod": {
            "$ref": "#/components/schemas/PaymentMethod"
          }
        },
        "additionalProperties": false
      },
      "PlanItem": {
        "type": "object",
        "properties": {
          "classId": {
            "type": "integer",
            "format": "int32"
          },
          "goal": {
            "type": "string",
            "nullable": true
          },
          "planId": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "plan": {
            "$ref": "#/components/schemas/Plan"
          },
          "class": {
            "$ref": "#/components/schemas/Class"
          }
        },
        "additionalProperties": false
      },
      "ProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": { }
      },
      "Purchase": {
        "type": "object",
        "properties": {
          "id": {
            "type": "integer",
            "format": "int32"
          },
          "city": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "street": {
            "type": "string",
            "nullable": true
          },
          "date": {
            "type": "string",
            "format": "date-time"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "total_price": {
            "type": "number",
            "format": "double"
          },
          "purchaseItems": {
            "type": "array",
            "items": {
              "$ref": "#/components/schemas/PurchaseItem"
            },
            "nullable": true
          },
          "paymentMethod": {
            "$ref": "#/components/schemas/PaymentMethod"
          }
        },
        "additionalProperties": false
      },
      "PurchaseDTO": {
        "type": "object",
        "properties": {
          "city": {
            "type": "string",
            "nullable": true
          },
          "country": {
            "type": "string",
            "nullable": true
          },
          "street": {
            "type": "string",
            "nullable": true
          },
          "totalPrice": {
            "type": "number",
            "format": "double"
          },
          "description": {
            "type": "string",
            "nullable": true
          },
          "paymentMethod": {
            "$ref": "#/components/schemas/PaymentMethod"
          }
        },
        "additionalProperties": false
      },
      "PurchaseItem": {
        "type": "object",
        "properties": {
          "itemId": {
            "type": "integer",
            "format": "int32"
          },
          "amount_bought": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number",
            "format": "double"
          },
          "purchaseId": {
            "type": "integer",
            "format": "int32"
          },
          "purchase": {
            "$ref": "#/components/schemas/Purchase"
          },
          "item": {
            "$ref": "#/components/schemas/Item"
          }
        },
        "additionalProperties": false
      },
      "PurchaseItemsDTO": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "brand": {
            "type": "string",
            "nullable": true
          },
          "quantity": {
            "type": "integer",
            "format": "int32"
          },
          "price": {
            "type": "number",
            "format": "double"
          }
        },
        "additionalProperties": false
      },
...
      "ValidationProblemDetails": {
        "type": "object",
        "properties": {
          "type": {
            "type": "string",
            "nullable": true
          },
          "title": {
            "type": "string",
            "nullable": true
          },
          "status": {
            "type": "integer",
            "format": "int32",
            "nullable": true
          },
          "detail": {
            "type": "string",
            "nullable": true
          },
          "instance": {
            "type": "string",
            "nullable": true
          },
          "errors": {
            "type": "object",
            "additionalProperties": {
              "type": "array",
              "items": {
                "type": "string"
              }
            },
            "nullable": true
          }
        },
        "additionalProperties": { }
      }
    }
  }
}